third_party/WebKit/LayoutTests/svg/dom/SVGAnimatedBoolean-expected.txt

This test checks the SVGAnimatedBoolean API - utilizing the preserveAlpha property of SVGFEConvolveMatrixElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial preserveAlpha value
PASS convElement.preserveAlpha.baseVal is false

Set value to true
PASS convElement.preserveAlpha.baseVal = true is true

Caching baseVal in local variable
PASS baseVal is true

Modify local baseVal variable to true
PASS baseVal = false is false

Assure that convElement.preserveAlpha has not been changed, but the local baseVal variable
PASS baseVal is false
PASS convElement.preserveAlpha.baseVal is true

Check assigning values of various types
PASS convElement.preserveAlpha.baseVal = convElement.preserveAlpha is convElement.preserveAlpha
PASS convElement.preserveAlpha.baseVal is true
PASS convElement.preserveAlpha.baseVal = null is null
PASS convElement.preserveAlpha.baseVal is false
PASS convElement.preserveAlpha.baseVal = 'aString' is 'aString'
PASS convElement.preserveAlpha.baseVal is true
PASS convElement.preserveAlpha.baseVal = convElement is convElement
PASS convElement.preserveAlpha.baseVal is true
PASS successfullyParsed is true

TEST COMPLETE
